r/membership: Fix import crash on incorrect ID
An incorrect two-part ID supplied to github_membership during import
results in a crash - this is because the import functionality is passing
this directly to the Read function of the resource, for which IDs are
programatically controlled and a panic there would be more telling of a
much more serious error worthy of panicking over.

This adds an internal validation function for two-part IDs, and wraps
the resource's import functionality to use it to validate incoming IDs
before passing them to Read.