Add teststep to check create default maintainer functionality.
The test checks firstly that a terraform config with create_default_maintainer
will add the token user as a member and maintainer of the team. It's important
to check as it's not managed be the read functions of github_team resource.

The updated config adds github_team_membership config to manage the user
membership created in the previous step.

Also added a function to test the team membership and team role of a user.