Fix compilation issues
After bumping to go-github v24, a couple of signatures have changed.

One of them, the `name` on `webhooks` has been removed. Previously, the
only valid name was `web`, and now, it has been removed from the object.

This commit is a breaking change on the provider, as it will not accept
more taking `name` on the webhook resources.