Refactor: cleanup usage of checkOrganization
Every instance where checkOrganization is called to ensure that
the provider has been configured with an organization name, the
code also requires the name itself, and obtains it later.

This patch changes the checkOrganization function to return the
name if it is present, and an error if it not, and renames it
getOrganization.

In addition there were some places where checkOrganization was
called but the organization name was not actually required, or
where it was obtained, included in a log message, but not used
in the resulting API call.

Signed-off-by: Kevin P. Fleming <kpfleming@bloomberg.net>