Refactor: cleanup usage of parseTwoPartID
This function is used in many resources to parse colon-separated
two-part identifier strings, and it emits an error message when
the string does not match the expected format. The error message
indicates that the format was expected to be "organization:name"
but this is only true for some resources, not all of them. As a
result users may get an error while attempting to import a resource
which gives them incorrect information about the identitier format.

This patch improves the function to accept a pair of label strings,
and it uses them when constructing the error message, so the message
will be tailored to the call site's desired format for the string.

Signed-off-by: Kevin P. Fleming <kpfleming@bloomberg.net>