Refactor: simplify API result handling
Every resource `Read` function has to process the results
of its primary GitHub API call, and there are a number of
possible results and actions to be taken.

This patch adds a function to parse the API results and return
an enum indicating the type of result the API returned; this
allows the calling code to use a `switch` statement to handle
each possible case, making the code more readable.

Signed-off-by: Kevin P. Fleming <kpfleming@bloomberg.net>