Convert github_team_membership to use user_id instead of username
This patch changes github_team_membership to use user_id to refer
to the user to be a team member, instead of the username. Usernames
are not static, and if the user changes their name in GitHub the
Terraform configuration will be broken. Using the user_id solves
this problem.

This patch includes a change to the schema for this resource,
and a StateUpgrader to upgrade existing state files to the new schema.

As a minor change, the validateTeamIDFunc schema validation function
was renamed to validateNumericIDFunc since it is now used for more
than just team_id validation.

Signed-off-by: Kevin P. Fleming <kpfleming@bloomberg.net>