Convert github_membership to use user_id instead of username
This patch changes github_membership to use user_id to refer
to the user to be an organization member, instead of the username. Usernames
are not static, and if the user changes their name in GitHub the
Terraform configuration will be broken. Using the user_id solves
this problem.

This patch includes a change to the schema for this resource,
and a StateUpgrader to upgrade existing state files to the new schema.

Signed-off-by: Kevin P. Fleming <kpfleming@bloomberg.net>