Add user_id support to data source github_user
This patch adds support for specifying user_id to the github_user
data source; the user can provide either user_id or username to look
up the user.

It also adds a boolean `minimal` argument to the data source, allowing
the user to avoid generation of the API calls needed to obtain the
GitHub user's GPG and SSH keys if they will not be used.

Signed-off-by: Kevin P. Fleming <kpfleming@bloomberg.net>