Prevent a panic when trying to use DismissalRestrictions
The recently updated `go-github` package updated the protections struct
a little by converting the `DismissalRestrictions` field from a struct
to a pointer to a struct. Hence this can now be `nil` which in turn
causes a panic when trying to use it as a value.

This PR added the change to the `go-github` package: https://github.com/google/go-github/pull/1288