Revert Config Changes To `github_branch_protection`
`repository` was previously removed in favour of `repository_id`.  We 
revert to `repository` here and transition `repository_id` to a computed 
value that is derived from a GraphQL lookup against the provided 
`repository`.

`branch` was previously removed in favour of `pattern`.  Similarly, 
`pattern` becomes computed and takes the value of `branch`.  `branch` 
now becomes deprecated for removal in the next major release.