Update resource_github_repository_file.go
If you have a repository with a high number of commits, you will quickly hit the Github API limit when trying to get all the commits for the repo. Adding the path parameter will limit the commits to only include the file in question, which will significantly reduce the number of API calls.