Add error message for importing branch protection
Add a useful error message when a user attempts to import a branch
protection rule that does not exist.

Before:

Error: nil entry in ImportState results. This is always a bug with
the resource that is being imported. Please report this as
a bug to Terraform.

After:

Error: Could not find a branch protection rule with the pattern 'master'.