Send a `User-Agent` in GitHub API requests with `terraform-provider-github` and the version
This updates the transport used to make requests to the GitHub
API to set a `User-Agent` request header with `terraform-provider-
github` and the current version (e.g. `terraform-provider-github/
4.26.1`).

Because this hooks in with the `RoundTrip` function, it can be
expected to overwrite `User-Agent` values set elsewhere (e.g.
as defaults or on a per-request basis by `go-github` or
`shurcooL/githubv4`).

This change does mean that someone will have to update the
`packageVersion` constant with the new version when it is
incremented. Perhaps there is a good way to avoid that!