[bug] Fix not detecting changes
- Do not rely on ETag to detect changes

github_repository does not detect changes in many of the arguments if
they are either done directly on GitHub or the resource is imported.
Here are some parameters which are affected (the list is not
comprehensive):

    allow_rebase_merge
    allow_squash_merge
    allow_update_branch
    delete_branch_on_merge

The reason for this behaviour seems to be related to how GitHub treats
ETag. Specifically ETag update is not triggered by the changes in the
parameters listed above.

See https://github.com/integrations/terraform-provider-github/issues/1446