Avoid GraphQL query to delete team maintainer upon creation
GraphQL doesn't yet support so called fine-grained PAT.
When creating a team, by default the current user is added as the team maintainer, and terraform tries to delete the maintainer unless the `create_default_maintainer` flag is set to true. The call to get the team mebers uses GrapQL resulting in a failure. The team gets created successfully but terraform apply fails with an error