Add build_type to datasource github_repository's pages
PR #1663 added the "build_type" attribute to the github_repository resource.
Since the corresponding data source uses the same code to parse the pages
endpoint data, any use of the github_repository datasource would lead to an
error, since the attribute did not exist there:

> Error: error setting pages: Invalid address to set: []string{"pages", "0", "build_type"}

This commit adds the missing attribute. The existing acceptance tests already
cover it, failed for me when running without this commit, and passed after.