Fixes accidental deletion of Repository resource from state
Currently when the user does not authenticate the provider, refreshing a GithubRepository resource drops it from the
state, which is unexpected and very confusing.

The root cause of this is: when AnonymousHTTPClient is used, owner == "" and this causes resourceGithubRepositoryRead to
issue requests to non-existent URLs such as https://github.com//myrepo and subsequently interpret 404 as a reason to
drop the resource from the state.

The suggested fix consults the resource data to infer the appropriate owner to use instead of the empty owner.