fix: Use ETag when reading `github_branch_default` resources.
This resource was missing saving and using the underlying resource's
ETag. This was discovered when we were hitting over-quota errors via
this resource.

The etag related implementation was copied from the `github_repository`
implementation.

Aside: There could be some refactoring to have these resources share
some common code. Such a refactoring seemed outside the scope of this
change.