Update security_and_analysis settings only when there are changes
The `resourceGithubRepositoryUpdate` function currently passes
the current state of security_and_analysis to the PATCH /repo
request. If the repo has an enforced security configuration,
this can result in the API returning a 422 response because
changing security settings is not allowed in this case.
Even though the security_and_analysis settings remain the same,
and only some other settings are changed, the request will still
fail.

To avoid the request failing when security_and_analysis remains
the same, we will only pass it in the request payload when changes
have been made.