(fix): add the handlers for empty values
This commit add the control of the allowed_repositories
and the allowed_actions for the actions_permission resources.
Without them the API calls will fail because the allowed_actions and
the allowed_repository are passed to the API with a null value.

The patch removes them from the HTTP request.