fix: ensure baseUrl is not already prefixed (#2616)
For self-hosted github (i.e with host different than https://api.github.com/) in the NewRESTClient func the "api/v3/" suffix is automatically added.
However in the `getInstallationAccessToken` the same check is done. This results in have a double suffix.
This fix ensure it is not added if already existing