Implement code security configuration resource for organziations
The implementation does not support a few attributes present in the
github rest api for code security configuration as these attributes
 are not present in the go-github library. Some (non-exhaustive) attributes include
code_scanning_delegated_alert_dismissal,
secret_scanning_delegated_alert_dismissal and
secret_scanning_generic_secrets.

Furthermore, I tried to implement default_for_new_repos as an attribute.
While it is possible to to set that attribute through a separate API
call, it is not currently possible to read the attribute through the
go-github sdk, meaning subsequent updates are not (currently) possible.
Hence, it was left out of this commit.

Fixes #2412