fix: allow missing types of rule bypasses (#2726)
Leave it to the API to determine if the values for actor_types are valid
of not. These change over time and we would need to carry a list in code
to verify this. Do not require an actor_id on rule bypasses. Things like
DeployKey do not have an id.

Signed-off-by: Johannes Wienke <languitar@semipol.de>
Co-authored-by: Randy Coburn <morfien101@gmail.com>
Co-authored-by: Nick Floyd <139819+nickfloyd@users.noreply.github.com>