Add handling of `AllowedMergeMethods`
It's a required field in the API and go-github doesn't use `omitempty`, so it submits `nil` if it isn't sent explicitly.

This change tries to keep it in the state, without having a configuration option for it (poor choice?)

And it defaults to all 3 available merge methods if it can't set something from the state.

Signed-off-by: Timo Sand <timo.sand@f-secure.com>